chore: moved schema
