+++ src/common/schema/campaigns.ts
@@ -1,0 +1,20 @@
+import z from 'zod';
+
+export const CAMPAIGN_VALIDATION_SCHEMA = z.object({
+  budget: z
+    .number()
+    .int()
+    .min(1000)
+    .max(100000)
+    .refine((value) => value % 1000 === 0, {
+      message: 'Budget must be divisible by 1000',
+    }),
+  duration: z
+    .number()
+    .int()
+    .min(1)
+    .max(30)
+    .refine((value) => value % 1 === 0, {
+      message: 'Duration must be a whole number',
+    }),
+});
